ai_agents/domain_cortex_agents/domain_cortex/requirements.txt
# Aetherion Domain Cortex - Python Dependencies

# Core dependencies
pyyaml>=6.0
httpx>=0.24.0
asyncio>=3.4.3

# Optional dependencies for advanced features
# numpy>=1.24.0  # For numerical computations in synthetic data
# pandas>=2.0.0  # For data analysis in child agents
